+++ Graphics/GraphicsEngineOpenGL/src/GLContextEmscripten.cpp
@@ -95,7 +95,7 @@ GLContext::GLContext(const EngineGLCreateInfo& InitAttribs, RENDER_DEVICE_TYPE& 
         LOG_ERROR_MESSAGE("Failed to enable SRGB framebuffers");
 
     DevType    = RENDER_DEVICE_TYPE_GLES;
-    APIVersion = Version{MajorVersion, MinorVersion};
+    APIVersion = Version{static_cast<Uint32>(MajorVersion), static_cast<Uint32>(MinorVersion)};
 }
 
 GLContext::~GLContext()
