+++ Graphics/GraphicsEngineOpenGL/src/GLContextLinux.cpp
@@ -80,7 +80,7 @@ GLContext::GLContext(const EngineGLCreateInfo& InitAttribs,
         LOG_ERROR_MESSAGE("Failed to enable SRGB framebuffers");
 
     DevType    = RENDER_DEVICE_TYPE_GL;
-    APIVersion = Version{MajorVersion, MinorVersion};
+    APIVersion = Version{static_cast<Uint32>(MajorVersion), static_cast<Uint32>(MinorVersion)};
 }
 
 GLContext::~GLContext()
